# cmd/api/docs/swagger.yaml
definitions:
  entity.Project:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      webhook_url:
        type: string
    type: object
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
    type: object
  request.AddProjectRequest:
    properties:
      name:
        maxLength: 255
        type: string
      slug:
        maxLength: 255
        type: string
      webhook_url:
        type: string
    required:
    - name
    - slug
    type: object
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  response.BaseResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  response.LoginData:
    properties:
      email:
        type: string
      id:
        type: integer
      token:
        type: string
    type: object
  response.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/response.LoginData'
      message:
        type: string
      status:
        type: integer
    type: object
  response.PaginationMetadata:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
  response.ProjectResponse:
    properties:
      message:
        type: string
      project:
        $ref: '#/definitions/entity.Project'
      status:
        type: integer
    type: object
  response.ProjectsResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMetadata'
      projects:
        items:
          $ref: '#/definitions/entity.Project'
        type: array
      status:
        type: integer
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is the documentation for the main e-commerce service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LogStream API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Perform login
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Perform register
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      summary: Register
      tags:
      - auth
  /projects/:
    get:
      consumes:
      - application/json
      description: Get All Project
      parameters:
      - in: query
        name: orderBy
        type: string
      - in: query
        name: page
        required: true
        type: integer
      - in: query
        name: pageSize
        required: true
        type: integer
      - in: query
        name: searchAll
        type: string
      - in: query
        name: searchField
        type: string
      - in: query
        name: searchValue
        type: string
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Project
      tags:
      - project
    post:
      consumes:
      - application/json
      description: Add new Project
      parameters:
      - description: Add Project request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AddProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Add Project
      tags:
      - project
  /projects/{id}:
    delete:
      description: Delete Project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Project
      tags:
      - project
    put:
      consumes:
      - application/json
      description: Update Project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add Project request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AddProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Project
      tags:
      - project
  /public/projects/{projectId}:
    get:
      consumes:
      - application/json
      description: Get All Project
      parameters:
      - in: query
        name: orderBy
        type: string
      - in: query
        name: page
        required: true
        type: integer
      - in: query
        name: pageSize
        required: true
        type: integer
      - in: query
        name: searchAll
        type: string
      - in: query
        name: searchField
        type: string
      - in: query
        name: searchValue
        type: string
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Project
      tags:
      - public
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"