# cmd/api/dto/request/register_request.go
package request

import (
	"encoding/json"
)

type RegisterRequest struct {	
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

func (r RegisterRequest) Marshal() ([]byte, error) {
	marshal, err := json.Marshal(r)

	if err != nil {
		return nil, err
	}

	return marshal, nil
}

func (r *RegisterRequest) Unmarshal(data []byte) error {
	return json.Unmarshal(data, &r)
}